% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cranhistory-package.R
\docType{package}
\name{cranhistory-package}
\alias{cranhistory}
\alias{cranhistory-package}
\title{cranhistory: Statistical Analysis of R Packages over Time}
\description{
Statistical analysis of R packages over time.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mpadge/cranhistory}
  \item Report bugs at \url{https://github.com/mpadge/cranhistory/issues}
}

}
\author{
\strong{Maintainer}: Mark Padgham \email{mark.padgham@email.com}

}
\keyword{internal}
